update setinterval carousel
update carousel
