+++ src/pages/Main/Main.scss
@@ -1,41 +1,7 @@
 @import '../../styles/variables.scss';
 
-.Main {
+.main {
   text-align: center;
-
-  .topCarousel {
-    margin-left: auto;
-    margin-right: auto;
-    max-width: 85.75rem;
-
-    .firstMain {
-      .firstPosition {
-        padding: 0 1.563rem;
-        width: 100%;
-      }
-
-      .Writing {
-        width: 100%;
-        height: 6rem;
-        margin-top: 1rem;
-        display: flex;
-        justify-content: space-around;
-
-        .firstWriting {
-          font-size: 2.5rem;
-          font-weight: bold;
-        }
-
-        .secondWriting {
-          text-align: end;
-
-          P {
-            font-size: 1.188rem;
-          }
-        }
-      }
-    }
-  }
 
   .goToCategorySite {
     display: flex;
@@ -48,7 +14,7 @@
       }
 
       p {
-        margin-top: 1rem;
+        margin: 1rem 0 0 0;
         font-size: 1.875rem;
         font-weight: bold;
       }
@@ -58,7 +24,7 @@
         border-radius: 0.25rem;
         height: 3.325rem;
         background: none;
-        margin-top: 0.75rem;
+        margin: 0.75rem 0 0 0;
         padding: 1rem;
       }
     }
