+++ src/pages/Main/components/ManProductsSlide/ManProductsSlide.scss
@@ -1,11 +1,9 @@
-@import '../../styles/variables.scss';
+@import '../../../../styles/variables.scss';
 
-.ManProductSlide {
+.manProductSlide {
   overflow: hidden;
   max-width: 85.75rem;
-  margin-left: auto;
-  margin-right: auto;
-  margin-bottom: 4rem;
+  margin: 0 auto 4rem auto;
   padding: 0 0.75rem;
   position: relative;
   text-align: center;
@@ -39,6 +37,27 @@
     right: 2rem;
     z-index: 1;
     cursor: pointer;
+  }
+
+  .containerDots {
+    position: absolute;
+    bottom: 4.5rem;
+    left: 50%;
+    transform: translateX(-50%);
+    display: flex;
+  }
+
+  .dot {
+    width: 1rem;
+    height: 1rem;
+    border-radius: 50%;
+    border: 0.188rem solid #f1f1f1;
+    margin: 0 5px;
+    background: #f1f1f1;
+  }
+
+  .dot.active {
+    background: rgb(32, 32, 32);
   }
 
   .goToShopNewProduct {
