+++ src/pages/Main/components/TopCarousel/TopCarousel.js
@@ -3,6 +3,7 @@ import './TopCarousel.scss';
 
 const TopCarousel = () => {
   const [sources, setSources] = useState([]);
+  const [currentSlide, setCurrentSlide] = useState(0);
 
   useEffect(() => {
     fetch('data/topCarousel.json', {
@@ -13,7 +14,21 @@ const TopCarousel = () => {
         setSources(data);
       });
   }, []);
-  const [currentSlide, setCurrentSlide] = useState(1);
+
+  useEffect(() => {
+    slideRef.current.style.transition = 'all 0.5s ease-in-out';
+    slideRef.current.style.transform = `translateX(-${currentSlide}00%)`;
+  }, [currentSlide]);
+
+  useEffect(() => {
+    const timer = setInterval(
+      () => setCurrentSlide(prevIndex => (prevIndex !== 1 ? prevIndex + 1 : 0)),
+      3000
+    );
+    return () => {
+      clearInterval(timer);
+    };
+  }, []);
 
   const slideRef = useRef(null);
 
@@ -21,18 +36,13 @@ const TopCarousel = () => {
     setCurrentSlide(index - 1);
   };
 
-  useEffect(() => {
-    slideRef.current.style.transition = 'all 0.5s ease-in-out';
-    slideRef.current.style.transform = `translateX(-${currentSlide}00%)`;
-  }, [currentSlide]);
-
   return (
     <div className="topCarousel">
       <div className="cardWrapper" ref={slideRef}>
         {sources.map(({ id, src, alt, leftside, rightside }) => (
           <div className="topCarouselCard" key={id}>
             <img className="firstPosition" src={src} alt={alt} />
-            <div className="Writing">
+            <div className="writing">
               <div className="firstWriting">{leftside}</div>
               <div className="secondWriting">
                 <p>{rightside}</p>
@@ -42,7 +52,7 @@ const TopCarousel = () => {
         ))}
       </div>
       <div className="container-dots">
-        {Array.from({ length: 2 }).map((item, index) => (
+        {Array.from({ length: 2 }).map((_, index) => (
           <div
             key={index}
             onClick={() => moveDot(index + 1)}
