+++ src/pages/Main/components/WomanProductsSlide/WomanProductsSlide.scss
@@ -1,11 +1,9 @@
-@import '../../styles/variables.scss';
+@import '../../../../styles/variables.scss';
 
-.WomanProductsSlide {
+.womanProductsSlide {
   overflow: hidden;
   max-width: 85.75rem;
-  margin-left: auto;
-  margin-right: auto;
-  margin-bottom: 4rem;
+  margin: 0 auto 4rem auto;
   padding: 0 0.75rem;
   position: relative;
   text-align: center;
@@ -41,19 +39,19 @@
     cursor: pointer;
   }
 
-  .container-dots {
+  .containerDots {
     position: absolute;
-    bottom: 10px;
+    bottom: 4.5rem;
     left: 50%;
     transform: translateX(-50%);
     display: flex;
   }
 
   .dot {
-    width: 20px;
-    height: 20px;
+    width: 1rem;
+    height: 1rem;
     border-radius: 50%;
-    border: 3px solid #f1f1f1;
+    border: 0.188rem solid #f1f1f1;
     margin: 0 5px;
     background: #f1f1f1;
   }
@@ -63,7 +61,7 @@
   }
 
   .goToShopNewProduct {
-    margin-top: 1.5rem;
+    margin: 3.2rem 0 0 0;
     padding: 1rem;
     font-weight: bold;
     border: 0.0625rem solid #000;
