+++ src/pages/Main/components/WomanProductsSlide/WomanSlideCard.js
@@ -4,7 +4,7 @@ import './WomanSlideCard.scss';
 const WomanSlideCard = ({ source }) => {
   const { src, productname, price } = source;
   return (
-    <div className="WomanSlideCard">
+    <div className="womanSlideCard">
       <img src={src} alt="lulu1" />
       <div className="productNamePrice">
         <div className="productName">{productname}</div>
